serverless/requirements.txt
aws-cdk-lib==2.186.0
constructs>=10.0.0,<11.0.0
boto3==1.38.23
botocore==1.38.23
jmespath==1.0.1
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
urllib3==2.4.0
